fft balling

+++ src/PolynomialMain.java
@@ -1,14 +1,14 @@
 public class PolynomialMain {
     public static void main(String[] args) {
-        int pass = 0;
-        int fail = 0;
-        polynomial p1 = new polynomial(new int[] {0,0,0,1});
-        polynomial p2 = new polynomial(new int[] {0,0,0,1});
+        double pass = 0;
+        double fail = 0;
+        polynomial p1 = new polynomial(new double[] {0,0,0,1});
+        polynomial p2 = new polynomial(new double[] {0,0,0,1});
 
         //test 1 add
         polynomial p3 = p1.add(p2);
         boolean same = true;
-        int[] p3_ideal = new int[] {0,0,0,2};
+        double[] p3_ideal = new double[] {0,0,0,2};
         for(int i = 0; i<p3.getVal().length; i++){
             if(p3_ideal[i]!=p3.getVal()[i]) same = false;
         }
@@ -18,7 +18,7 @@ public class PolynomialMain {
         //test 2 subtract
         polynomial p4 = p1.subtract(p2);
         same = true;
-        int[] p4_ideal = new int[] {0,0,0,0};
+        double[] p4_ideal = new double[] {0,0,0,0};
         for(int i = 0; i<p4.getVal().length; i++){
             if(p4_ideal[i]!=p4.getVal()[i]) same = false;
         }
@@ -28,7 +28,7 @@ public class PolynomialMain {
         //test 3 multiply
         polynomial p5 = p1.multiply(p2);
         same = true;
-        int[] p5_ideal = new int[] {0,0,0,0,0,0,1};
+        double[] p5_ideal = new double[] {0,0,0,0,0,0,1};
         for(int i = 0; i<p5.getVal().length; i++){
             if(p5_ideal[i]!=p5.getVal()[i]) same = false;
             //System.out.print(" "+ p5.getVal()[i] + " ");
@@ -41,11 +41,11 @@ public class PolynomialMain {
          */
 
         //test 4 add
-        polynomial p11 = new polynomial(new int[] {0,0,1,1}); // 0,0,1,1
-        polynomial p12 = new polynomial(new int[] {0,0,2,1});//0,2,2,0
+        polynomial p11 = new polynomial(new double[] {0,0,1,1}); // 0,0,1,1
+        polynomial p12 = new polynomial(new double[] {0,0,2,1});//0,2,2,0
         polynomial p13 = p11.add(p12);
         same = true;
-        int[] p13_ideal = new int[] {0,0,3,2};
+        double[] p13_ideal = new double[] {0,0,3,2};
         for(int i = 0; i<p13.getVal().length; i++){
             if(p13_ideal[i]!=p13.getVal()[i]) same = false;
         }
@@ -55,7 +55,7 @@ public class PolynomialMain {
         //test 5 subtract
         polynomial p14 = p11.subtract(p12);
         same = true;
-        int[] p14_ideal = new int[] {0,0,-1,0};
+        double[] p14_ideal = new double[] {0,0,-1,0};
         for(int i = 0; i<p14.getVal().length; i++){
             if(p14_ideal[i]!=p14.getVal()[i]) same = false;
         }
@@ -65,7 +65,7 @@ public class PolynomialMain {
         //test 6 multiply
         polynomial p15 = p11.multiply(p12);
         same = true;
-        int[] p15_ideal = new int[] {0,0,0,0,2,3,1};
+        double[] p15_ideal = new double[] {0,0,0,0,2,3,1};
         for(int i = 0; i<p15.getVal().length; i++){
             if(p15_ideal[i]!=p15.getVal()[i]) same = false;
             //System.out.print(" "+ p5.getVal()[i] + " ");
@@ -78,15 +78,15 @@ public class PolynomialMain {
          */
 
         //test 7 add
-        polynomial p21 = new polynomial(new int[] {0,1,1,1});
-        polynomial p22 = new polynomial(new int[] {0,1,2,1});
+        polynomial p21 = new polynomial(new double[] {0,1,1,1});
+        polynomial p22 = new polynomial(new double[] {0,1,2,1});
         // 0, 0, 0, 0, 1, 1, 1
         // 0, 0, 0, 2, 2, 2, 0
         // 0, 0, 1, 1, 1, 0, 0
         // 0, 0, 1, 3, 4, 3, 1
         polynomial p23 = p21.add(p22);
         same = true;
-        int[] p23_ideal = new int[] {0,2,3,2};
+        double[] p23_ideal = new double[] {0,2,3,2};
         for(int i = 0; i<p23.getVal().length; i++){
             if(p23_ideal[i]!=p23.getVal()[i]) same = false;
         }
@@ -96,7 +96,7 @@ public class PolynomialMain {
         //test 8 subtract
         polynomial p24 = p21.subtract(p22);
         same = true;
-        int[] p24_ideal = new int[] {0,0,-1,0};
+        double[] p24_ideal = new double[] {0,0,-1,0};
         for(int i = 0; i<p24.getVal().length; i++){
             if(p24_ideal[i]!=p24.getVal()[i]) same = false;
         }
@@ -106,7 +106,7 @@ public class PolynomialMain {
         //test 9 multiply
         polynomial p25 = p21.multiply(p22);
         same = true;
-        int[] p25_ideal = new int[] {0, 0, 1, 3, 4, 3, 1};
+        double[] p25_ideal = new double[] {0, 0, 1, 3, 4, 3, 1};
         for(int i = 0; i<p25.getVal().length; i++){
             if(p25_ideal[i]!=p25.getVal()[i]) same = false;
             //System.out.print(" "+ p5.getVal()[i] + " ");
@@ -119,8 +119,8 @@ public class PolynomialMain {
          */
 
         //test 10 add
-        polynomial p2g1 = new polynomial(new int[] {0,1,0,1});
-        polynomial p2g2 = new polynomial(new int[] {0,1,2,1});
+        polynomial p2g1 = new polynomial(new double[] {0,1,0,1});
+        polynomial p2g2 = new polynomial(new double[] {0,1,2,1});
         // 0, 0, 0, 0, 1, 0, 1
         // 0, 0, 0, 2, 0, 2, 0
         // 0, 0, 1, 0, 1, 0, 0
@@ -128,7 +128,7 @@ public class PolynomialMain {
         // 0, 0, 1, 2, 2, 2, 1
         polynomial p2g3 = p2g1.add(p2g2);
         same = true;
-        int[] p2g3_ideal = new int[] {0,2,2,2};
+        double[] p2g3_ideal = new double[] {0,2,2,2};
         for(int i = 0; i<p2g3.getVal().length; i++){
             if(p2g3_ideal[i]!=p2g3.getVal()[i]) same = false;
         }
@@ -138,7 +138,7 @@ public class PolynomialMain {
         //test 11 subtract
         polynomial p2g4 = p2g1.subtract(p2g2);
         same = true;
-        int[] p2g4_ideal = new int[] {0,0,-2,0};
+        double[] p2g4_ideal = new double[] {0,0,-2,0};
         for(int i = 0; i<p2g4.getVal().length; i++){
             if(p2g4_ideal[i]!=p2g4.getVal()[i]) same = false;
         }
@@ -148,7 +148,7 @@ public class PolynomialMain {
         //test 12 multiply
         polynomial p2g5 = p2g1.multiply(p2g2);
         same = true;
-        int[] p2g5_ideal = new int[] {0, 0, 1, 2, 2, 2, 1};
+        double[] p2g5_ideal = new double[] {0, 0, 1, 2, 2, 2, 1};
         for(int i = 0; i<p2g5.getVal().length; i++){
             if(p2g5_ideal[i]!=p2g5.getVal()[i]) same = false;
             //System.out.print(" "+ p2g5.getVal()[i] + " ");
